# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered user wrapped in a consistent response."
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session role, name, and email for the frontend."
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/guides": {
      "get": {
        "operationId": "GuidesController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List tour guides."
          }
        },
        "tags": [
          "guides"
        ]
      },
      "post": {
        "operationId": "GuidesController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGuideDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create tour guide."
          }
        },
        "tags": [
          "guides"
        ]
      }
    },
    "/guides/{id}": {
      "get": {
        "operationId": "GuidesController_findOne",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find tour guide."
          }
        },
        "tags": [
          "guides"
        ]
      },
      "patch": {
        "operationId": "GuidesController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGuideDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update tour guide."
          }
        },
        "tags": [
          "guides"
        ]
      },
      "delete": {
        "operationId": "GuidesController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete tour guide."
          }
        },
        "tags": [
          "guides"
        ]
      }
    },
...
    "/issues": {
      "get": {
        "operationId": "IssuesController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "issues"
        ]
      },
      "post": {
        "operationId": "IssuesController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report an issue to support."
          }
        },
        "tags": [
          "issues"
        ]
      }
    },
    "/issues/{id}": {
      "get": {
        "operationId": "IssuesController_findOne",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "issues"
        ]
      },
      "patch": {
        "operationId": "IssuesController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIssueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "issues"
        ]
      },
      "delete": {
        "operationId": "IssuesController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "issues"
        ]
      }
    },
    "/issues/{id}/resolve": {
      "post": {
        "operationId": "IssuesController_resolve",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveIssueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resolve an issue and notify every trip member."
          }
        },
        "tags": [
          "issues"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dream Destination Review 4 API",
    "description": "NestJS in-memory REST API with RBAC header authorization for traveler, travel partner, tour guide, vendor, support, and super user portals.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role header required on every API. Use Super User, Traveler, Travel Partner, Tour Guide, Vendor, or Support Executive."
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sarah Johnson"
          },
          "email": {
            "type": "string",
            "example": "sarah@example.com"
          },
          "phone": {
            "type": "string",
            "example": "9876543210"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "role": {
            "type": "string",
            "example": "Traveler"
          }
        },
        "required": [
          "name",
          "email",
          "phone",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "traveler@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "role": {
            "type": "string",
            "example": "Traveler"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateGuideDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Arun Kumar"
          },
          "languages": {
            "type": "string",
            "example": "English, Hindi"
          },
          "experience": {
            "type": "string",
            "example": "8 years"
          },
          "rating": {
            "type": "number",
            "example": 4.8
          }
        },
        "required": [
          "name",
          "languages",
          "experience"
        ]
      },
      "UpdateGuideDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Arun Kumar"
          },
          "languages": {
            "type": "string",
            "example": "English, Hindi"
          },
          "experience": {
            "type": "string",
            "example": "8 years"
          },
          "rating": {
            "type": "number",
            "example": 4.8
          }
        }
      },
...
      "CreateIssueDto": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "string",
            "example": "TRIP-101"
          },
          "title": {
            "type": "string",
            "example": "Hotel reservation issue"
          },
          "description": {
            "type": "string",
            "example": "Hotel could not find the booking."
          },
          "type": {
            "type": "string",
            "example": "Accommodation"
          },
          "priority": {
            "type": "string",
            "example": "High"
          },
          "reportedBy": {
            "type": "string",
            "example": "Sarah Johnson"
          },
          "reporterRole": {
            "type": "string",
            "example": "Traveler"
          }
        },
        "required": [
          "tripId",
          "title",
          "description"
        ]
      },
      "UpdateIssueDto": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "string",
            "example": "TRIP-101"
          },
          "title": {
            "type": "string",
            "example": "Hotel reservation issue"
          },
          "description": {
            "type": "string",
            "example": "Hotel could not find the booking."
          },
          "type": {
            "type": "string",
            "example": "Accommodation"
          },
          "priority": {
            "type": "string",
            "example": "High"
          },
          "reportedBy": {
            "type": "string",
            "example": "Sarah Johnson"
          },
          "reporterRole": {
            "type": "string",
            "example": "Traveler"
          }
        }
      },
      "ResolveIssueDto": {
        "type": "object",
        "properties": {
          "resolution": {
            "type": "string",
            "example": "Support contacted hotel and confirmed reservation."
          }
        },
        "required": [
          "resolution"
        ]
      },
...
    }
  }
}